requirements.txt
websocket-client
pandas
requests
pytz
python-telegram-bot==20.7
